requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
fastapi==0.115.12
filelock==3.13.1
flatbuffers==25.2.10
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.3
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
onnx==1.18.0
onnxruntime==1.22.0
packaging==25.0
pillow==11.0.0
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==70.2.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
python-dotenv==1.0.1

# Database and caching for real data integration
asyncpg>=0.28.0
redis>=4.5.0
sqlalchemy>=2.0.0
alembic>=1.10.0

# HTTP client and async support
aiohttp>=3.8.0
httpx>=0.24.0

# External integrations
psycopg2-binary>=2.9.0

# Monitoring and logging
prometheus-client>=0.16.0
structlog>=22.0.0

# Security
cryptography>=40.0.0
passlib>=1.7.0
python-jose>=3.3.0

# Data processing
pytz>=2023.0
python-dateutil>=2.8.0